group.cpp: Use get_device_defined to search for devices
When adding devices to a group using a wildcard only exported devices
are added, this is due to GroupElementFactory using
`get_device_exported` to search for devices matching the wildcard.
Now we are using `get_device_defined` we match unexported devices too.
This is more consistent as it is already possible to add unexported
devices when adding them individually.