Make starting a Device Server on macOS from any directory on the system possible
Correct the __darwin__ specific code in void DServer::class_factory() to
make it possible to start a device server on macOS from any directory on the
system. With the old code the DS could not be started unless the DS executable
was run from the same DS executable's directory. Naturally this was very rarely
the case and hence this really needed to be fixed in order to work for any
directory.

General modifications:
- Introduce const std::string& Util::get_ds_unmodified_exec_name() which returns
  argv[0] of a Device Server executable. argv[0] is stored in the private
  member ds_unmodified_exec_name in Util.

Modifications for macOS only:
- Call get_ds_unmodified_exec_name from void DServer::class_factory() to
  to retrieve the device server's executable name and resolve the full path
  of it with the help of realpath from the std. C library.
- Use the full path when dlopen the DS executable.
- Significantly improve error messages. When dlsym cannot find the
  mangled class factory symbol in a Device Server file that was opened with
  dlopen, then the error message needs to tell the user what went wrong and if
  possible why.
- Add documentation of why something is done.
- Overwriting of the union member needs to happen after initialisation.