Fix bug in WAttribute::set_write_value for DevState RW attributes (#963)
WAttribute::set_write_value no longer throws an API_IncompatibleArgumentType
exception when trying to set the write value of a DevState RW attribute.
This bug was due to a wrong assumption that std::is_enum<T> is returning True
only for the DevEnum attributes types but DevState is also seen as an enum by
std::is_enum<T>.
As a consequence, the wrong method was called (the one written for the DevEnum).