tango_config.h: Introduce TANGO_REV_EXP
The use cases here is quite obscure. We need to export a DLL interface
when compiling a device server *outside* and *against* cppTango. We
therefore call the macro reverse export.

This is just how it is done right now.

Both cases are slightly different as DServer::class_factory looks at
TANGO_HAS_DLL but DeviceClass does not. Always exporting it does not hurt
though.