configure/CMakeLists.txt: Prefer CMAKE_CURRENT_XXX variables
Only these variables work correctly when the cppTango project is included
into another cmake project using add_subdirectory.

All occurences found with

git grep -P 'CMAKE_(BINARY|SOURCE)_DIR'

The match in configure/cmake_pch_linux.cmake was not handled as PCH is
currently broken.

The patch in tests also drops the "tests" folder as this is a
separate project which already lives in "tests" relative to the repository
root.