Fix Precompiled Header Feature (#989)
A regression was introduced when merging the header file sepearation MR (!952)
when using -DTANGO_USE_PCH=on CMake option.
The tango.h include file given with -include option was not located in a
directory containing the corresponding tango.h.gch precompiled header, resulting
in a compilation time as long as when using -DTANGO_USE_PCH=OFF CMake option.
Please refer to !526 MR to get details on how the precompiled headers are
supposed to be implemented in cppTango.