appveyor.yml: Build and install turbojpeg like other dependencies on Windows
Build a debug version of turbojpeg when we are building
cppTango in debug mode.

We set -DCMAKE_DEBUG_POSTFIX=d when building turbojpeg so that the names
of the debug libraries don't collide with the release versions.

An option JPEG_DEBUG_POSTFIX has been added to the cppTango
CMakeLists.txt so that it can be linked against debug vesions of
turbojpeg which haven't been built this way.

Like other dependencies, turbojpeg is only installed if
TANGO_INSTALL_DEPENDENCIES=ON and TANGO_USE_JPEG=ON.

When installing tubojpeg we install:
- all header files to include\
- all static libraries and import libraries to lib\
- all .dll files and .pdb files to bin\

Note: This installs more than we strictly speaking need for cppTango,
    however, we may as well provide all of turbojpeg just in case the
    user wants it.