CMakeLists.txt: Add define with git describe output
In that way package maintainers and users know the exact version.

The EXISTS check is done here instead of the more obvious IS_DIRECTORY
as that also works when cppTango is a submodule and .git is a file.

In case something went wrong fetching the revision, we set it to
unknown.