CMakeLists.txt: Distinguish released versions from development versions
We do that by having a SUFFIX_VERSION which is automatically set by git.
For tagged versions this will be empty and it will be dev for all other
cases.

For the development versions we will then have

lrwxrwxrwx 1 firma firma   14 10. Nov 22:54 libtango.so -> libtango.so.94
lrwxrwxrwx 1 firma firma   20 10. Nov 22:54 libtango.so.94 -> libtango.so.9.4.0dev
-rwxr-xr-x 1 firma firma 8,9M 10. Nov 22:54 libtango.so.9.4.0dev

and for the release versions

lrwxrwxrwx 1 firma firma   14 10. Nov 23:00 libtango.so -> libtango.so.94
lrwxrwxrwx 1 firma firma   17 10. Nov 23:00 libtango.so.94 -> libtango.so.9.4.0
-rwxr-xr-x 1 firma firma 8,9M 10. Nov 23:00 libtango.so.9.4.0