ZmqEventSupplier: Add the published ORB endPoint to the list of alternate endpoints
In 8e13daf5 (- Change the event connection way in case of DS running on
hosts with several NIC., 2015-06-15) the support for several ZQM endpoints
was added. The idea was that the server has a list of all endpoints which
are available (ZmqSupplier: alternate_h_endpoint/alternate_e_endpoint).
This list can be fetched by the client via ZmqEventSubscriptionChangeCmd.
And now the client can iterate through all available end points in
connect_event_channel/check_zmq_endpoint and use a working one.

While the code works for the stated design goal of multiple network cards.

It fails with device servers in containers. The reason is that the
container instance itself does *not* know at which outside hostname it
will be reachable. But there is currently also no way to pass that
information in.

For CORBA transport the designers envisoned that issue and added
ORBendPointPublish. This is part of the IOR and allows the server to tell
the client how to reach it [1].

We now also read that from the environment and add that to the list of
alternate endpoints for both sockets.

[1]: 6.6.3 Endpoint publishing in
     https://omniorb.net/omni42/omniORB/omniORB006.html#chap%3Aconnections