configure/CMakeLists.txt: Check omniORB version
We only support a few versions (as demonstrated by cppTango#1013 and
cppTango#760) of omniORB.  We should ensure that a supported version has
been found at cmake configuration time.

If omniORB is found by pkg-config we use the version information from
the .pc file, otherwise the version is extracted from
omniORB4/acconfig.h using cmake regular expressions.

If the version does not fall in an acceptable range we raise an error.
If we cannot determine the version information we issue warnings but do
not stop the build.

This check is performed for both Windows and Linux builds.