ZmqEventConsumer::connect_event_channel: Handle hostnames in list of endpoints
Since forever we expected only IPv4 addresses in the list of endpoints and
alternate endpoints from ZmqEventSubscriptionChange.

But we would actually like to use hostnames for containerized server and
client.

To support that we now try to resolve the endpoint if it does not look
like an IPv4 address. And if we could resolve it we pass it to
check_zmq_endpoint for verification.

Doing it in this order allows to save some CPU cycles in the common case
of IPv4 addresses in endpoints.