configure/CMakeLists.txt: Fetch the omniidl executable from pkg-config
When users use a custom pkg-config location with PKG_CONFIG_PATH we have
to use the omniidl binary provided there.

In addition we now also don't search for the omniidl binary anymore with
find_program for the pkg-config case.

For the TANGO_OMNI_BASE case we only search in the passed location and not in
system folders.