CMake: Allow to supply dependent libraries via pkg-config only
Users who prefer to use the more standard approch of using
PKG_CONFIG_PATH for finding custom library locations instead of our
TANGO_XXX_BASE variables are currently out of luck as we don't support that.

For fixing that we need to add the library paths to the global include
directories, target include directories don't work with cmake 3.7.

And also add the standard cflags/include directories for all other
targets.