configure/CMakeLists.txt: Cleanup ZMQ variable handling
We now always set the pgk-config variable ZMQ_PKG_LIBRARY_DIRS. This can
be used to simplify our try_compile checks and also ensures that a
user supplied PKG_CONFIG_PATH is honoured.

One complication is that at least libzmq version 4.0.5 only sets
ZMQ_PKG_LIBDIR, so we have to check for that as well.