utils: Make argv/argc handling more future proof
Since forever the code assumed that when it sees a CLI argument which
expects a parameter like -ORBEndpoint that the user also supplied one.

In the code this was done a la

for (int i = 2;i < argc;i++)
  if(strcmp(param.c_str(),argv[i]) == 0)
   do something with argv[i + 1];

Now this seems to be a plain out-of-bounds access. But documentation in [1] states

  The size of the array pointed to by argv is at least argc+1, and the last
  element, argv[argc], is guaranteed to be a null pointer.

so it is actually safe to do. But all code currently constructs a
std::string from that, and doing that from a nullptr is UB.

So let's warn the user in this case.

[1]: https://en.cppreference.com/w/c/language/main_function