cxx_attr_write: Add an error desc test for ...
... some API_IncompatibleAttrDataType errors.

This test anticipates a future commit where we will improve the error
message produced by the following function overload:

    WAttribute::check_written_value(const Tango::AttrValUnion &, ...)

This overload is used when a client writes to an attribute.  The test
attempts to write a value of the wrong type to a device server via a
DeviceProxy object and checks that the caught error contains
information about the expected type and the found type.