cxx_attr_misc:  Add another error desc test for...
... more API_IncompatibleAttrDataType errors.

This test anticipates a future commit where we will improve the error
message produced by the following function overload:

    WAttribute::check_written_value(const CORBA::Any &, ...)

This overload is used when an attribute is written to directly by the
server.  The test uses a new IOAttrThrowEx flag (io_init_wattr_except)
to instruct the DevTest device to throw an exception when the
IOInitWAttr is called.  When this flag is set the server will attempt to
write an attribute with a value of the wrong type.  The test then checks
that the caught error contains information about the expected type and
the found type.

IOInitWAttr has been made a friend to DevTest so that it has access to
the new IOAttrThrowEx flag.

The overload can also be called when an old client (pre Tango IDL v4)
tries to write to an attribute.  This code path is not tested.