configure/find_and_check_omniidl.cmake: Do not run check when cross compiling
`try_run` command fails when we are cross compiling unless
CMAKE_CROSSCOMPILING_EMULATOR is set.  We should detect this situation
and avoid calling `try_run`.

The best we can do is just hope that omniidl doesn't have the bug and
skip the check.  We check the omniORB4 version and avoid the 4.2.1
release where we know this bug occurs, so this check isn't going to be
catching much anyway.

Removed the now unnecessary setting of `OMNIIDL_TEST_RUN` when cross
compiling and the advice to do this in INSTALL.md.