detail::parse_hostname_from_CORBA_URI: Allow empty hostnames
The refactoring in f388207a (Util::check_end_point_specified: Factor out
host/ip parsing, 2023-01-17) wanted to be really clever and disallowed
empty hostnames.

That is wrong as per [1]

> Either the host or port, or both can be left empty.

The fallout is that it was impossible to use end point specifications like

giop:tcp::5000

which are common for the database DS and, as it turns out, also widely
used in other cases.

Fix that by allowing and empty hostname again and adapt the tests.

[1]: https://omniorb.sourceforge.io/omni42/omniORB/omniORB006.html#chap%3Aconnections