cmake: Allow to specify omniidl search folder
On MacOSX 13.2.1 M1 when using the homebrew package of omniORB, its
installation folder for the headers is separate from omniidl.

This breaks our logic with TANGO_OMNI_BASE which expects that everything is
below the given folder.

Fix that by introducing TANGO_OMNIIDL_PATH which allows specify a search hint
for omniidl.