abi-api-compliance-check: Identify user for git merge
In 017830bd (abi-api-compliance-check: Update for new merge strategy,
2023-04-18) we changed the abi-api-compliance-check job to create the new
branch by merging the MR branch into main.

This fails on the CI when we need to do a real merge (i.e. not a
fast-forward) because git requires us to identify ourselves to commit
anything.  Even though we specify `--no-commit` git still requires this.

This was missed previously as the MR was only tested on the CI when we
only needed a fast-forward merge.