WAttribute::check_written_value: Improve...
...error messages for CORBA::Any overload.

Add information about the type we found when we get a type mismatch and
report type names rather than CmdArgType enum values. The error messages
are intended to be the same as the AttrValUnion overload, except when
the type is unknown where information relevant for each overload is
added to aid in debugging.

We determine which type we have found inside the CORBA::Any using
CORBA::TypeCode::equivalent rather than CORBA::TypeCode::equal
as this better reflects the semantics of the CORBA::Any::operator>>=
used to do the extraction.

Also, remove the now unused _throw_incompatible_exception(long)
overload.