tests: Remove Windows/Linux hostname manipulation
It is unclear in what circumstances this manipulation is required and we
do not test it in our CI.  As the comment suggests that this is to do
with compatibility between Windows and Linux and the hostnames come from
`cmake_host_system_information` it seems more appropriate that this
should be done in CMake if it is actually required.

Also, this manipulation assumes that here is a '.' in the hostname and
will crash if this is not the case.

For now, we remove this from the C++ tests with a plan to add something
to the CMakeLists.txt file if, in the future, it turns out that the
hostnames are actually not compatible between operating systems.