cppapi: Enable compatibility parameter for omniORB C++ < 4.3.0
From [1]:

If true (the default), when an Exception is inserted in an Any, it is transmit-
ted as the TypeCode, followed by the Exception repository id, followed by the
Exception data members. This is the format used by other CORBA implementa-
tions, including omniORBpy. Versions of omniORB for C++ prior to 4.3 did not
send the repository id after the TypeCode. If you require interoperability with
earlier omniORB versions that send or receive exceptions inside Anys, set this
parameter to false.

And we do want to have interoperability with omniORB C++ 4.2.x and
earlier.

[1]: https://omniorb.sourceforge.io/omni43/omniORB.pdf