.gitlab-ci.yml: Print backtraces if we find core dumps
Printing a full backtrace for a crashing program in the CI output will
often allow us to track down an issue without having to download and
interpret the core file locally.

We add debug information to release builds so that we can get useful
stack traces for those builds too.