.gitlab-ci.yml: Fix core dump generation and collection
Docker requires us to set the core ulimit in order for crashing
processes to generate core dumps. Presumably, this means core dump
generation has been broken since we moved to running the tests inside
docker containers.

Device servers and test programs are run from build/tests directory, so
this is where core files are generated.  We update the artifact pattern
to correctly collect these core files.  We can determine which test the
correspond to which core dump based on the PID in the core dump file
name, which is printed by the `run_with_fixtures.sh` script.