setup_database.sh: Increase timeout for mysql server
The CI tests have been sporadically failing on the SKAO runners due to
the mysql container taking a longer than 30 seconds to start when the
runners are under load.

Increasing the timeout avoids this situation.  We must use a new
release of the tango-db image with an increased timeout waiting for the
mysql server port.