Use target_precompile_headers from cmake
This commit changes the way the tango.h header is precompiled. It uses
the target_precompile_headers cmake function available since v3.16. The
precompiled header is reused as much as possible to avoid recompilation.

The cmake minimum version is set to 3.18 which is above 3.16 and match
the minimum version decided in issue #975