.gitlab-ci.yml: Use docker image for abi-api-compliance check
Our CI docker images all use a default user tango which is different from
the user doing the gitlab checkout in CI. And therefore our calls to git
worktree add in ci/check-ABI-API-compliance.sh fail as the user in the
docker image can not write into the .git folder.

The workaround is to skip the git checkout and do it ourselves.