Switch to Gitlab CI on Windows
As appveyor support is not that good on gitlab, and appveyor also only
gives us one concurrent job, we decided to move to gitlab CI for windows
as well.

We currently do use the shared runners from gitlab, but plan to switch to
our own runners in the future.

The compilation approach has not changed, we still use our own
pre-packaged dependencies.

The switch from appveyor to the gitlab shared runners also required us to
upgrade from Visual Studio 2017 to Visual Studio 2019. This is not a ABI
issue as both are compatible [1].

[1]:https://learn.microsoft.com/en-us/cpp/porting/binary-compat-2015-2017?view=msvc-170