tests/multi_thread_sighandler.cpp:  Get compiling on Windows
As we don't run the tests on Windows CI currently, for now we just get the
test to compile.  This is done by mostly #ifdef'ing out code, making the
test useless on Windows.

The test relies on fork, which does not exist on Windows, so getting
this running will require some work.