find_and_check_omniidl.cmake: Raise error on check failure
In ffc053a6 (configure/CMakeLists.txt: Fetch the omniidl executable from
pkg-config, 2023-01-31) the check for a valid omniidl binary was move
from after the inclusion of `find_and_check_omniidl.cmake` to before the
inclusion.

`find_and_check_omniidl.cmake` was relying on the old order to
actually raise an error when a problem occurred.  It was setting
`OMNIIDL` to `FALSE` and letting the later check for `OMNIIDL` send a
`FATAL_ERROR` message.

When the order was switch, CMake only emits a warning at configuration
time if the omniidl check fails.

Let's have the failure be generated in the script directly to avoid this
ordering dependence.