Quote string values to file database correctly
The logic deciding whether to quote string values written into the file
database was slightly flawed: while it considered values with spaces as
needing to be quoted, it didn't consider values with *quotes* (which
were escaped) as needing quoting. For example, the strings `hi` , `hi `,
`"hi"` and `"hi "`would end up being written as `hi`, `"hi "`, `\"hi\"`
and `"\"hi \""`` respectively -- with the third one then failing to be
read correctly.

This commit fixes this situation by considering values with spaces *or*
double quotes as needing quoting when writing then to the file database.
Given the current state of the code, I also took the opportunity to
extract this writing of string values into a small utility function to
remove duplicated code, and also rewrote the escape_double_quote
function for easier reading.

This addresses #1162.