Remove TANGO_GIT_REVISION macro
Note that this is a small backwards-compatibility breaking change: while
the result is still the same when using the macro value in the code, it
cannot be used during pre-processing anymore. The macro was introduced
only in 921b074b though, so isn't used anywhere internally, nor is it
used in pytango, nor is it mentioned in the online documentation either,
so remove it is a safe operation.

Although this is a small change, it has a big effect: it means that upon
cmake reconfiguration the tango_const.h header is not re-generated
anymore. This is a core header of the project, and hence its
regeneration triggers a (full? it certainly seems so) recompilation of
the project.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>