CheckEnumCompatibleType: Allow enums with integer as underlying type as well
The default underlying type for enumerations for MSVC is int [1]. And our
enforcement of having either short or unsigned int as underlying type in
the previous commit made our windows CI fail.

We now can either do two things:
- Reject enumerations with underlying type int on MSCV. That would be a
  breaking change for users.
- Allow enumerations with underlying type int on all platforms.

The author of this commit chooses the second option as that seems most
compatible. We also don't run into storage space issues as we already
allow unsigned int and that has the same size as int.

[1]: https://learn.microsoft.com/en-us/cpp/build/reference/zc-enumtypes?view=msvc-170