configure/check_omniidl.cmake: Don't require static omniORB libraries
We have a report that latest cppTango 9.5.0-rc1 does not build on
conda-forge. The issue there is that conda-forge does build static
libraries, so our omniidl test fails.

The solution here is to check if the static target exists, and only then
use it, otherwise we fallback to trying the dynamic library.