CheckEnumCompatibleType: Restrict underlying enum types to just short/DevShort
The RFC specification says in [1]

> Values of DevEnum DataType SHALL be transported as values of DevShort.

So our current support for int or unsigned int is not covered by an RFC
requirement. We also don't transport integers larger than short faithfully
over the network. And since 9.4 pogo also uses short as underlying type.

So at todays cppTango developers meeting it was decided that the support
for other types than short is erroneous and should be removed.

While at it we also switch from short to DevShort, both are equivalent,
which is also verified by a static assertion.

[1]: https://tango-controls.readthedocs.io/projects/rfc/en/latest/9/DataTypes.html#enumeration