Regenerate git description on each commit
The mechanism currently in place (introduced in 519b2da0 and bdd1c072)
to generate the git description is slightly flawed, as it retrieves this
information only when cmake goes through its configuration phase,
without the information being regenerated when the HEAD has moved to a
different reference (e.g., when changing branches, adding/removing
commits, etc).

This commit improves on this situation by scanning the contents of the
.git/HEAD file and adding that file, and whatever if points to, to the
list of cmake configuration dependencies. With these changes, a cmake
reconfiguration is issued whenever the HEAD changes. Since this now
results only on `git_revision.cpp` being regenerated the cost is much
smaller than it used to be, and the information contained in the binary
is always correct.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>