server/ntservice.cpp: Remove default case in control code handler
In 63a5e379 (TANGO_THROW_ON_DEFAULT: Introduce and use it, 2022-07-26)
we added a default case to throw an exception in the Win32 control
handler function. We cannot actually throw an exception here, as it can
never be caught.  Also, asserting (introduced in 11637ca6 (assert.h: Add
TANGO_ASSERT_ON_DEFAULT, 2023-10-13)) is not ideal as it is Windows
which sends the control codes to the process so we cannot statically
guarantee that it will not fire. Instead Microsoft recommends just
ignoring unhandled codes[1].

[1] https://learn.microsoft.com/en-us/windows/win32/services/writing-a-control-handler-function