configure/CMakeLists.txt:  Use find modules in TangoCMakeModules
The FindomniORB4.cmake from TangoCMakeModules is missing the
`/D_WIN32_WINNT=0x0400` compared to the version deleted by this commit.
However, unlike what is suggested in the omniORB README.win32.txt the
web documentation[1] does not say that we have to set this pneumonic to
use omniORB.

We already set `_WIN32_WINNT=0x0500` in `tango_const.h` if it is
undefined, so this commit does define it from CMake and let's
`tango_const.h` handle it.

[1] https://omniorb.sourceforge.io/omni43/omniORB/omniORB001.html