tests/conf_devtest.cpp: Clean up print_changes overloads
The `std::string&` overload cannot accept `const char*` parameters as
a temporary `std::string` cannot be passed as a non-const lvalue
reference.

We do not modified the arguments so it is fine to mark them as `const`.

With `const std::string&` parameters the `const char*` overload is
unneeded and now we can accept a mix of `const char*` and `std::string`
arguments.