tests: Catch exceptions thrown by TangoPrinter::{start,kill}_server
In the previous commit, we updated these functions to throw an exception
when the `system()` call failed.  Here we update call sites to catch
this exception to avoid `std::terminate()` being called, `abort()`ing
the test.

In tests we just assert that these functions do not throw (or that they
do throw if they are expected to fail).

In setup/teardown contexts, we catch the error and print a message but
carry on.  This preservers the previous control flow, where the error
from the `system()` call is ignored, except that now we get an error
message telling up about it.