CmpTst: Use unique filenames based on PID
Now that we are running tests against IDL5 and IDL6 servers, we can be
running a test executable multiple times at once when running `ctest
-j<n>` (i.e. once for against the IDL5 and once against the IDL6
server). This causes issues for any test using CmpTst as the test will
always open the same files.  If two copies of these tests run at the
same time, they end up clobbering each other.

Using unique filenames for output files and patched reference files
avoids this clobbering.

Now that the output filename is unique per run of the test, we also have
to ensure that if the filename occurs in the reference file it is
patched correctly.