bdd/test_attr_manip.cpp: Add class to run test against
To run the test against the device server, run the following to launch
the device server:

BddServer test -nodb -ORBendPoint giop:tcp:127.0.0.1:10000 -dlist "AttrManipDev::attr_manip/test/1"

Then run BddTests to run the test against this server.  This does not
require a Tango database.

Removed the test from ctest as it assumes the device is reachable with
"#dbase=no" so will not work with DevTest.