ci/check-ABI-API-compliance.sh: Initialise submodules
The TangoCMakeModules git submodule was added to the repository in
b6a89543 (TangoCMakeModules: Add it, 2023-11-28) and the
ci/check-ABI-API-compliance.sh script was not updated to handle this.

When creating branches with `git worktree`, submodules are not checked
out.  There is no "recurse" flag for `git worktree add` so the best we
can do is initialise the submodules after we have created the worktree.

Note: git-worktree(1) recommends not using worktrees and submodules
together because worktrees are "experimental" and "support for
submodules is incomplete".  However, this use case seems to work (at
least with git 2.39.2 provided by debian bookworm).