.clang-tidy: Demote performance-avoid-endl to warning
Although I agree with the advice to avoid `endl` in favor of `'\n'`,
changing `endl` to `'\n'` can result in a stream not being flushed at
the same point.

This might have some subtle consequences so I don't think it is safe to
apply the auto-fix without manual auditing.  Safer is just to demote
this to a warning.