server/blackbox.cpp: Avoid copy initialization
This satisfies the performance-unnecessary-copy-initialization
clang-tidy check.  The check was auto-fixed in 6b84cb19 (clang-tidy:
Auto-fix performance-unnecessary-copy-initialization, 2023-09-02) and
has been failing since c3a65891 (add support for new IDL6 client
identification data struct in the client API - #1193, 2023-11-17).

The regression was missed because the clang-tidies warnings-as-errors
was broken in 9a7eb485 (.clang-tidy: Suppress failing warnings,
2023-09-18).