bdd/CMakeLists.txt: Generate a single executable
Catch2 is generally provided as a static library on Linux
distributions.  Linking this into a shared object is not possible unless
the libCatch2.a has been compiled with `-fPIC`.  This is not the case on
e.g. Ubuntu, so before this change would not be able to compile these
tests without building Catch2 yourself.

We look at `argv[0]` to determine if we are running the tests or a Tango
device server, this keeps the CLI for the tests expected by
`catch_discover_tests`.