tests/CMakeLists.txt: Readd _REENTRANT define
Since the introduction of opentelemetry with dependencies we get a warning
like

cc1plus: warning: /home/firma/devel/cppTango/build-main/tests/CMakeFiles/conf_devtest.dir/cmake_pch.hxx.gch: not used because `_REENTRANT' is defined [-Winvalid-pch]

when compiling.

The exact cause of the warning is unknown, but re-adding the define which
was removed in ef9a2187 (CMake: Remove _REENTRANT define, 2023-11-16)
fixes the issue so let's do that.