.windows-gitlab-ci.yml: Don't run the release job for dev tags
We recently changed our release process, see db77020b (RELEASING.md:
Update to tag start of a release's development, 2024-01-26).

We now add a tag ending on `-dev` at the start of the release cycle. For
these tags we don't want to have a release created.

See [1] for the source of the !~ expression.

[1]: https://docs.gitlab.com/ee/ci/jobs/job_control.html#common-if-clauses-for-rules