Logging::init: Accept database as a pointer
In Utils::effective_job we sometimes don't have a database object when
we call Logging::init.  With Logging::init taking a reference we need to
dereference a nullptr to construct the reference.  This is undefined
behaviour even though we do not access the db reference in this case as
`use_db` is false.

We cannot use the truthiness of the database point to replace `use_db`
because sometimes the database object is a file database which should
not be used for logging.