catch2/utils/callback_mock.h: Add CallbackMock class template
This allows us to write sequential tests interacting with the Tango
event system without having to sleep for events.

The CallbackMock here is based on the class template that Michal Liszcz
wrote in !588.

For older releases of glibc, libpthreads.so is a separate library which
we have to link against explicitly when using C++11 threading features.

We use the CMake target Thread::Thread handles this for us.