catch2/utils/utils.cpp: Move ApiUtil::cleanup to end of tests
ApiUtil::cleanup is very slow when working with event tests, as cleaning
up the ZMQ threads takes ~0.5s.

We only actually need to do this just before the test process exits and
not in between each test case.  All the functionality we are testing
should work for "long lived" Tango clients as well as "short lived"
Tango clients.