Connection::Cb_ReadAttr_Request: Write errors into correct list
When we have an error from device attribute reading from IDLv3 onwards we
want to pass on the errors to the callback function via
AttrReadEvent::errors.

By copying to the DevErrorList errors object we ensure that
AttrReadEvent::err is set correctly in its ctor.

This branch is not executed for IDLv1/v2, there we take the "the reply is
an exception" else branch.