zmq_event_subscription_change: Use create_full_event_name to...
... construct ev_name, rather than custom code.

`create_full_event_name` handles the nodb case correctly, which was not
done in the custom code.

`ev_name` is used to store an event count in the `event_cptr` map.
In the nodb case, when we went to push an event we could not find the
count as `push_event` uses `create_full_event_name` correctly.

The missing count in the `event_cptr` map, results in an event being
sent to the client with a count of 0.

To work around a ZMQ bug, the client is supposed to skip the first event
(with count == 1) in some circumstances.  This skip was not happening in
the nodb case as the count was always 0, so the client failed to detect the
first event.  This resulted in an extra event being pushed to the user
callback in the nodb case.

With this change the extra event is no long pushed to the user.