ZmqEventConsumer::push_zmq_event: Avoid integer overflow
ds_ctr can be zero if there is a bug in the server resulting in it not
finding the event in its count map.  When this is the case we get the
following UBSAN failure:

    /builds/[MASKED]/cppTango/src/client/zmqeventconsumer.cpp:2018:37: runtime error: unsigned integer overflow: 0 - 1 cannot be represented in type 'unsigned int'

Avoiding the subtraction if ds_ctr == 0 does not change the behaviour:

If evt_cb.ctr > 0 then

    - with this change:  missed_event == -evt_cb.ctr
    - previously: missed_event == 1 (strictly speaking it is UB, but in
      practice this is what we get)

In both cases we take the final else branch.