Catch2Tests: Add --only-idl flag to only run against a single idl version
When debugging an issue, it is often unnecessary to run the tests
multiple times against servers of different versions.  This flags allows
you to avoid that.

In the case where you specify `--only-idl` outside the range for a given
test it is skipped.  This requires Catch 3.3.0.

We could fail the test inside which would avoid the need for Catch
3.3.0, but I do not think it is a big deal to increase the catch
version. Skipping the test means we do not confuse this with real
failures.