ci/windows-build.ps1: Add and use Invoke-NativeCommand
Now that we are running the build job from a powershell script, the CI
runner is not running commands directly which means it is not checking
error codes.

Unfortunately, powershell does not have a `set -e` type option to check
error codes for commands, so in order to get the script to fail when a
command returns a non-zero error code we need to check $LASTEXITCODE and
throw an exception.