tests/catch2_alarm_event.cpp: Refactor tests to use a generator
Before this commit the tests did run as intended because sequential Catch2 WHEN
blocks are not executed in the same test partial case (i.e. we spawn a
new TestServer for each of these WHEN blocks).

Using the GENERATE_REF macro and explicitly setting the initial value for
each test partial case we actually run the test that was intended in
each case.