Attribute::log_quality: Handle user defined ALARM/WARNING qualities
Attribute::quality can be set to ATTR_WARNING/ATTR_ALARM here by the
user calls `set_value(value, date, quality)`.  In this case, the alarm
bit field is empty, so we fail the default case assertion.

This fixes the "Scenario: Alarm events can be pushed from code manually"
test which was aborting due to the assertion failure before this change.