catch2/utils/utils.cpp: Fix log filename transformation on Windows
`std::string_view::iterator` is not guaranteed to be convertible to
`const char *` by the standard, and in fact on Windows it is not.  This
leads to the following error from MSVC when we try to initialise `end`:

  C:\GitLab-Runner\builds\tango-controls\cppTango\tests\catch2\utils\utils.cpp(260,50): error C2440: '=': cannot convert from 'std::_String_view_iterator<_Traits>' to 'const char *' [C:\GitLab-Runner\builds\tango-controls\cppTango\build\tests\Catch2Tests.vcxproj]

This has been broken since 5fe007dc (catch2/utils/test_server.cpp:
Factor out filename_from_test_case_name, 2024-03-13) and was missed
because the Windows CI job was broken.