FileDatabase: Add limited support for free object properties
Support includes reading the properties, of the "CtrlSystem" free
object, from the file and writing them back again. If no free object
properties are found in the file, there is no change in behaviour.

During file parsing the FREE "domain" was already being checked for (and
ignored), so this should not introduce any change of behaviour for
existing files.

This allows us to test the new AutoAlarmOnChangeEvent property from the
new catch2 tests.