TangoTest::AutoDeviceClass: Add TANGO_TEST_ENABLED_CLASSES env var
This allows registering only a subset of the device classes in the
device server at startup time.

This is needed as when we use the file database, it expects to find an
entry for each registered device class.  By avoiding registering
device classes we aren't interested in, we don't need to include them
all in the file database.

It might also help with startup times of the device servers.