Attribute::fire_change_event: Call fire_alarm_event if AutoAlarmOnChangeEvent
Before this commit, AutoAlarmOnChangeEvent was being implemented in
`detect_and_push_change_event`, having it call
`do_detect_and_push_alarm_event` as appropriate.

This is not the correct place to implement this feature because:

1. `detect_and_push_change_event` is only called if there are change
   event subscribers.  This means if there are only alarm event
   subscribers they will not receive any alarm events

2. `detect_and_push_change_event` is called by the polling thread.  This
   means that (if there are change event subscribers) any alarm event
   subscriber will receive two alarm events.

Instead we should be calling `fire_alarm_event` in `fire_change_event`
as this is only used by `push_change_event` and is always called,
regardless of whether there are any change event subscribers.